for-ref: Xilinx vendor build, Workaround for Ubuntu 24.04
Ubuntu 24.04 (and probibly later) has turned off access to user namespaces
by default.  This prevents bitbake from running.

This can be re-enabled on a per application basis but I could not find a
pre-crafted profile for bitbake.

I did find this global workaround that disables this restriction for
the current boot cycle.